#1: Separated read and write variant of the bucket.
