+++ src/main/java/com/exasol/bucketfs/AbstractBucketBuilder.java
@@ -1,0 +1,5 @@
+package com.exasol.bucketfs;
+
+public abstract class AbstractBucketBuilder<T extends AbstractBucketBuilder<T>> {
+    protected abstract T self();
+}
