+++ src/main/java/com/exasol/bucketfs/Bucket.java
@@ -1,127 +1,19 @@
 package com.exasol.bucketfs;
 
-import static com.exasol.containers.ExasolContainerConstants.SUPPORTED_ARCHIVE_EXTENSIONS;
-
-import java.io.IOException;
 import java.io.InputStream;
-import java.net.*;
-import java.net.http.*;
-import java.net.http.HttpRequest.BodyPublisher;
-import java.net.http.HttpRequest.BodyPublishers;
-import java.net.http.HttpResponse.BodyHandlers;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Instant;
-import java.time.temporal.ChronoField;
-import java.util.*;
 import java.util.concurrent.TimeoutException;
 import java.util.function.Supplier;
-import java.util.logging.Logger;
 
-/**
- * An abstraction for a bucket inside Exasol's BucketFS.
- */
-public class Bucket {
-    private static final Logger LOGGER = Logger.getLogger(Bucket.class.getName());
-    private static final String BUCKET_ROOT = "";
-    private static final long BUCKET_SYNC_TIMEOUT_IN_MILLISECONDS = 60000;
-    private static final long FILE_SYNC_POLLING_DELAY_IN_MILLISECONDS = 200;
-    private final String bucketFsName;
-    private final String bucketName;
-    private final String ipAddress;
-    private final int port;
-    private final String readPassword;
-    private final String writePassword;
-    private final HttpClient client = HttpClient.newBuilder().build();
-    private final BucketFsMonitor monitor;
-    private final Map<String, Instant> uploadHistory = new HashMap<>();
-
-    private Bucket(final Builder builder) {
-        this.bucketFsName = builder.bucketFsName;
-        this.bucketName = builder.bucketName;
-        this.ipAddress = builder.ipAddress;
-        this.port = builder.port;
-        this.readPassword = builder.readPassword;
-        this.writePassword = builder.writePassword;
-        this.monitor = builder.monitor;
-    }
-
-    /**
-     * @return name of the BucketFS filesystem this bucket belongs to
-     */
-    public String getBucketFsName() {
-        return this.bucketFsName;
-    }
-
-    /**
-     * @return name of the bucket
-     */
-    public String getBucketName() {
-        return this.bucketName;
-    }
-
-    /**
-     * Get the fully qualified name of the bucket.
-     *
-     * @return fully qualified name consisting of service name and bucket name
-     */
-    public String getFullyQualifiedBucketName() {
-        return this.bucketFsName + BucketConstants.PATH_SEPARATOR + this.bucketName;
-    }
-
-    /**
-     * Get the read password for the bucket.
-     *
-     * @return read password
-     */
-    public String getReadPassword() {
-        return this.readPassword;
-    }
+public interface Bucket {
 
     /**
      * Get the write password for the bucket.
      *
      * @return write password.
      */
-    public String getWritePassword() {
-        return this.writePassword;
-    }
-
-    /**
-     * List the contents of a bucket.
-     *
-     * @return bucket contents
-     * @throws BucketAccessException if the contents are not accessible or the path is invalid
-     * @throws InterruptedException  if the list request was interrupted
-     */
-    public List<String> listContents() throws BucketAccessException, InterruptedException {
-        return listContents(BUCKET_ROOT);
-    }
-
-    /**
-     * List the contents of a path inside a bucket.
-     *
-     * @param path relative path from the bucket root
-     * @return list of file system entries
-     * @throws BucketAccessException if the contents are not accessible or the path is invalid
-     * @throws InterruptedException  if the list request was interrupted
-     */
-    // [impl->dsn~bucket-lists-its-contents~1]
-    public List<String> listContents(final String path) throws BucketAccessException, InterruptedException {
-        final URI uri = createPublicReadURI(BUCKET_ROOT);
-        LOGGER.fine(() -> "Listing contents of bucket under URI \"" + uri + "\"");
-        try {
-            final HttpRequest request = HttpRequest.newBuilder(uri).build();
-            final HttpResponse<String> response = this.client.send(request, BodyHandlers.ofString());
-            if (response.statusCode() == HttpURLConnection.HTTP_OK) {
-                return parseContentListResponseBody(response, removeLeadingSlash(path));
-            } else {
-                throw new BucketAccessException("Unable to list contents of bucket.", response.statusCode(), uri);
-            }
-        } catch (final IOException exception) {
-            throw new BucketAccessException("Unable to list contents of bucket.", uri, exception);
-        }
-    }
+    String getWritePassword();
 
     /**
      * Check whether the file with the given path exists in this bucket.
@@ -134,58 +26,15 @@ public class Bucket {
      *                               the synchronization is missing.
      */
     // [impl->dsn~validating-bucketfs-object-synchronization-via-the-bucketfs-log~1]
-    public boolean isObjectSynchronized(final String pathInBucket, final Instant afterUTC)
-            throws InterruptedException, BucketAccessException {
-        if (hasSynchronizationMonitor()) {
-            throw new BucketAccessException("Unable to determine whether \"" + pathInBucket + "\" in \""
-                    + getFullyQualifiedBucketName()
-                    + "\" is synchronized. Please register a synchronization monitor when creating the bucket object.");
-        } else {
-            return this.monitor.isObjectSynchronized(this, pathInBucket, afterUTC);
-        }
-    }
+    boolean isObjectSynchronized(String pathInBucket, Instant afterUTC)
+            throws InterruptedException, BucketAccessException;
 
     /***
      * Check if a synchronization monitor is registered for this Bucket.
      *
      * @return {@code true} if a monitor is available
      */
-    public boolean hasSynchronizationMonitor() {
-        return this.monitor == null;
-    }
-
-    private String removeLeadingSlash(final String path) {
-        if (path.startsWith("/")) {
-            return path.substring(1);
-        } else {
-            return path;
-        }
-    }
-
-    private URI createPublicReadURI(final String pathInBucket) {
-        return URI.create("http://" + this.ipAddress + ":" + this.port + "/" + this.bucketName + "/"
-                + removeLeadingSlash(pathInBucket));
-    }
-
-    private List<String> parseContentListResponseBody(final HttpResponse<String> response, final String path) {
-        final String[] items = response.body().split("\\s+");
-        final List<String> contents = new ArrayList<>(items.length);
-        for (final String item : items) {
-            final String relativeItem = removeLeadingSlash(item);
-            if (relativeItem.startsWith(path)) {
-                contents.add(extractFirstPathComponent(relativeItem.substring(path.length(), relativeItem.length())));
-            }
-        }
-        return contents;
-    }
-
-    private String extractFirstPathComponent(final String path) {
-        if (path.contains("/")) {
-            return path.substring(0, path.indexOf('/'));
-        } else {
-            return path;
-        }
-    }
+    boolean hasSynchronizationMonitor();
 
     /**
      * Upload a file to the bucket.
@@ -204,10 +53,8 @@ public class Bucket {
      * @throws BucketAccessException if the file cannot be uploaded to the given URI
      */
     // [impl->dsn~uploading-to-bucket~1]
-    public void uploadFile(final Path localPath, final String pathInBucket)
-            throws InterruptedException, BucketAccessException, TimeoutException {
-        uploadFile(localPath, pathInBucket, true);
-    }
+    void uploadFile(Path localPath, String pathInBucket)
+            throws InterruptedException, BucketAccessException, TimeoutException;
 
     /**
      * Upload a file to the bucket and block the call until it is synchronized.
@@ -228,109 +75,14 @@ public class Bucket {
      * @throws TimeoutException      if synchronization takes too long
      */
     // [impl->dsn~uploading-to-bucket~1]
-    public void uploadFile(final Path localPath, final String pathInBucket, final boolean blocking)
-            throws InterruptedException, BucketAccessException, TimeoutException {
-        final String extendedPathInBucket = extendPathInBucketDownToFilename(localPath, pathInBucket);
-        try {
-            uploadContent(BodyPublishers.ofFile(localPath), extendedPathInBucket, "file " + extendedPathInBucket,
-                    blocking);
-        } catch (final IOException exception) {
-            throw new BucketAccessException("I/O failed to open file \"" + localPath + "\" for upload to BucketFS.",
-                    exception);
-        }
-    }
-
-    // Wait some time between uploads of the same file so we can distinguish the upload success logs for detecting
-    // upload success.
-    // [impl->dsn~bucketfs-object-overwrite-throttle~1]
-    private void delayRepeatedUploadToSamePath(final String extendedPathInBucket) throws InterruptedException {
-        if (this.uploadHistory.containsKey(extendedPathInBucket)) {
-            final Instant lastUploadAt = this.uploadHistory.get(extendedPathInBucket).with(ChronoField.NANO_OF_SECOND,
-                    0);
-            final Instant now = Instant.now();
-            if (!now.isAfter(lastUploadAt.plusSeconds(1))) {
-                final long delayInMillis = 1000L - (now.getNano() / 1000000L);
-                LOGGER.fine(() -> "Delaying upload to \"" + extendedPathInBucket + "\" for " + delayInMillis + " ms");
-                Thread.sleep(delayInMillis);
-            }
-        } else {
-            LOGGER.fine(() -> "No previous uploads to \"" + extendedPathInBucket
-                    + "\" recorded in upload history. No upload delay required.");
-        }
-    }
-
-    private String extendPathInBucketDownToFilename(final Path localPath, final String pathInBucket) {
-        return pathInBucket.endsWith("/") ? pathInBucket + localPath.getFileName() : pathInBucket;
-    }
-
-    private void uploadContent(final BodyPublisher bodyPublisher, final String pathInBucket,
-            final String contentDescription, final boolean blocking)
-            throws InterruptedException, BucketAccessException, TimeoutException {
-        if (blocking) {
-            delayRepeatedUploadToSamePath(pathInBucket);
-        }
-        final long millisSinceEpochBeforeUpload = System.currentTimeMillis();
-        uploadContentNonBlocking(bodyPublisher, pathInBucket, contentDescription);
-        if (blocking) {
-            waitForFileToBeSynchronized(pathInBucket, millisSinceEpochBeforeUpload);
-        }
-    }
-
-    private void uploadContentNonBlocking(final BodyPublisher bodyPublisher, final String pathInBucket,
-            final String contentDescription) throws InterruptedException, BucketAccessException {
-        final URI uri = createWriteUri(pathInBucket);
-        LOGGER.info(() -> "Uploading \"" + contentDescription + "\" to bucket \"" + this.bucketFsName + "/"
-                + this.bucketName + "\": \"" + uri + "\"");
-        try {
-            final int statusCode = httpPut(uri, bodyPublisher);
-            if (statusCode != HttpURLConnection.HTTP_OK) {
-                LOGGER.severe(
-                        () -> statusCode + ": Failed to upload \"" + contentDescription + "\" to \"" + uri + "\"");
-                throw new BucketAccessException("Unable to upload file \"" + contentDescription + "\" to ", statusCode,
-                        uri);
-            }
-        } catch (final IOException exception) {
-            throw new BucketAccessException("I/O error trying to upload \"" + contentDescription + "\" to ", uri,
-                    exception);
-        }
-        LOGGER.fine(() -> "Successfully uploaded to \"" + uri + "\"");
-        recordUploadInHistory(pathInBucket);
-    }
-
-    private void recordUploadInHistory(final String pathInBucket) {
-        final Instant now = Instant.now();
-        LOGGER.fine(() -> "Recorded upload to \"" + pathInBucket + "\" at " + now + " in upload history");
-        this.uploadHistory.put(pathInBucket, now);
-    }
-
-    private URI createWriteUri(final String pathInBucket) throws BucketAccessException {
-        try {
-            return new URI("http", null, this.ipAddress, this.port, "/" + this.bucketName + "/" + pathInBucket, null,
-                    null).normalize();
-        } catch (final URISyntaxException exception) {
-            throw new BucketAccessException("Unable to create write URI.", exception);
-        }
-    }
-
-    private int httpPut(final URI uri, final BodyPublisher bodyPublisher) throws IOException, InterruptedException {
-        final HttpRequest request = HttpRequest.newBuilder(uri) //
-                .PUT(bodyPublisher) //
-                .header("Authorization", encodeBasicAuth(true)) //
-                .build();
-        final HttpResponse<String> response = this.client.send(request, BodyHandlers.ofString());
-        return response.statusCode();
-    }
-
-    private String encodeBasicAuth(final boolean write) {
-        return "Basic " + Base64.getEncoder() //
-                .encodeToString((write ? ("w:" + this.writePassword) : ("r:" + this.readPassword)).getBytes());
-    }
+    void uploadFile(Path localPath, String pathInBucket, boolean blocking)
+            throws InterruptedException, BucketAccessException, TimeoutException;
 
     /**
      * Upload the contents of a string to the bucket.
      * <p>
      * This method is intended for writing small objects in BucketFS dynamically like for example configuration files.
-     * For large payload use {@link Bucket#uploadFile(Path, String)} instead.
+     * For large payload use {@link WriteEnabledBucket#uploadFile(Path, String)} instead.
      * </p>
      * <p>
      * This call blocks until the uploaded file is synchronized in BucketFs or a timeout occurs.
@@ -343,16 +95,14 @@ public class Bucket {
      * @throws TimeoutException      if synchronization takes too long
      */
     // [impl->dsn~uploading-strings-to-bucket~1]
-    public void uploadStringContent(final String content, final String pathInBucket)
-            throws InterruptedException, BucketAccessException, TimeoutException {
-        uploadStringContent(content, pathInBucket, true);
-    }
+    void uploadStringContent(String content, String pathInBucket)
+            throws InterruptedException, BucketAccessException, TimeoutException;
 
     /**
      * Upload the contents of a string to the bucket.
      * <p>
      * This method is intended for writing small objects in BucketFS dynamically like for example configuration files.
-     * For large payload use {@link Bucket#uploadFile(Path, String)} instead.
+     * For large payload use {@link WriteEnabledBucket#uploadFile(Path, String)} instead.
      * </p>
      * <p>
      * When blocking is enabled, this call waits until either the uploaded file is synchronized or a timeout occurred.
@@ -366,12 +116,8 @@ public class Bucket {
      * @throws BucketAccessException if the file cannot be uploaded to the given URI
      * @throws TimeoutException      if synchronization takes too long
      */
-    public void uploadStringContent(final String content, final String pathInBucket, final boolean blocking)
-            throws InterruptedException, BucketAccessException, TimeoutException {
-        final String excerpt = (content.length() > 20) ? content.substring(0, 20) + "..." : content;
-        final String description = "text " + excerpt;
-        uploadContent(BodyPublishers.ofString(content), pathInBucket, description, blocking);
-    }
+    void uploadStringContent(String content, String pathInBucket, boolean blocking)
+            throws InterruptedException, BucketAccessException, TimeoutException;
 
     /**
      * Upload the contents of an input stream to the bucket.
@@ -386,10 +132,8 @@ public class Bucket {
      * @throws TimeoutException      if synchronization takes too long
      */
     // [impl->dsn~uploading-input-stream-to-bucket~1]
-    public void uploadInputStream(final Supplier<InputStream> inputStreamSupplier, final String pathInBucket)
-            throws InterruptedException, BucketAccessException, TimeoutException {
-        uploadInputStream(inputStreamSupplier, pathInBucket, true);
-    }
+    void uploadInputStream(Supplier<InputStream> inputStreamSupplier, String pathInBucket)
+            throws InterruptedException, BucketAccessException, TimeoutException;
 
     /**
      * Upload the contents of an input stream to the bucket.
@@ -406,185 +150,7 @@ public class Bucket {
      * @throws TimeoutException      if synchronization takes too long
      */
     // [impl->dsn~dsn~uploading-input-stream-to-bucket~1]
-    public void uploadInputStream(final Supplier<InputStream> inputStreamSupplier, final String pathInBucket,
-            final boolean blocking) throws InterruptedException, BucketAccessException, TimeoutException {
-        uploadContentNonBlocking(BodyPublishers.ofInputStream(inputStreamSupplier), pathInBucket, "input stream");
-    }
+    void uploadInputStream(Supplier<InputStream> inputStreamSupplier, String pathInBucket, boolean blocking)
+            throws InterruptedException, BucketAccessException, TimeoutException;
 
-    /**
-     * Determine if the object at the given path has one of the archive formats that BucketFS can automatically expand.
-     *
-     * @param pathInBucket path to the object in the bucket
-     * @return {@code true} if the object has an archive format that BucketFS can automatically expand.
-     */
-    public static boolean isSupportedArchiveFormat(final String pathInBucket) {
-        for (final String extension : SUPPORTED_ARCHIVE_EXTENSIONS) {
-            if (pathInBucket.endsWith(extension)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    // [impl->dsn~waiting-until-archive-extracted~1]
-    // [impl->dsn~waiting-until-file-appears-in-target-directory~1]
-    private void waitForFileToBeSynchronized(final String pathInBucket, final long millisSinceEpochBeforeUpload)
-            throws InterruptedException, TimeoutException, BucketAccessException {
-        final long expiry = millisSinceEpochBeforeUpload + BUCKET_SYNC_TIMEOUT_IN_MILLISECONDS;
-        final Instant afterUtc = Instant.ofEpochMilli(millisSinceEpochBeforeUpload);
-        while (System.currentTimeMillis() < expiry) {
-            if (this.monitor.isObjectSynchronized(this, pathInBucket, afterUtc)) {
-                return;
-            }
-            Thread.sleep(FILE_SYNC_POLLING_DELAY_IN_MILLISECONDS);
-        }
-        final String message = "Timeout waiting for object \"" + pathInBucket + "\" to be synchronized in bucket \""
-                + getFullyQualifiedBucketName() + "\" after " + afterUtc + ".";
-        LOGGER.severe(() -> message);
-        throw new TimeoutException(message);
-    }
-
-    /**
-     * Download a file from a bucket to a local filesystem.
-     *
-     * @param pathInBucket path of the file in BucketFS
-     * @param localPath    local path the file is downloaded to
-     * @throws InterruptedException  if the file download was interrupted
-     * @throws BucketAccessException if the local file does not exist or is not accessible or if the download failed
-     */
-    // [impl->dsn~downloading-a-file-from-a-bucket~1]
-    public void downloadFile(final String pathInBucket, final Path localPath)
-            throws InterruptedException, BucketAccessException {
-        final URI uri = createPublicReadURI(pathInBucket);
-        LOGGER.fine(() -> "Downloading  file from bucket \"" + this.bucketFsName + "/" + this.bucketName + "\": \""
-                + uri + "\" to \"" + localPath + "\"");
-        try {
-            final int statusCode = httpGet(uri, localPath);
-            if (statusCode != HttpURLConnection.HTTP_OK) {
-                LOGGER.severe(() -> statusCode + ": Failed to download \"" + uri + "\" to file \"" + localPath + "\"");
-                throw new BucketAccessException("Unable to downolad file \"" + localPath + "\" from ", statusCode, uri);
-            }
-        } catch (final IOException exception) {
-            throw new BucketAccessException("Unable to upload file \"" + localPath + "\" from ", uri, exception);
-        }
-        LOGGER.fine(() -> "Successfully downloaded file to \"" + localPath + "\"");
-    }
-
-    private int httpGet(final URI uri, final Path localPath) throws IOException, InterruptedException {
-        final HttpRequest request = HttpRequest.newBuilder(uri) //
-                .GET() //
-                .header("Authorization", encodeBasicAuth(true)) //
-                .build();
-        final HttpResponse<String> response = this.client.send(request, BodyHandlers.ofString());
-        Files.write(localPath, response.body().getBytes());
-        return response.statusCode();
-    }
-
-    /**
-     * Create builder for a {@link Bucket}.
-     *
-     * @return builder
-     */
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    /**
-     * Builder for {@link Bucket} objects.
-     */
-    public static class Builder {
-        private String bucketFsName;
-        private String bucketName;
-        private String ipAddress;
-        private int port;
-        private String readPassword;
-        private String writePassword;
-        private BucketFsMonitor monitor;
-
-        /**
-         * Set monitor for this bucket.
-         *
-         * @param monitor synchronization montior
-         * @return Builder instance for fluent programming
-         */
-        public Builder monitor(final BucketFsMonitor monitor) {
-            this.monitor = monitor;
-            return this;
-        }
-
-        /**
-         * Set the filesystem name.
-         *
-         * @param bucketFsName name of the BucketFS filesystem
-         * @return Builder instance for fluent programming
-         */
-        public Builder serviceName(final String bucketFsName) {
-            this.bucketFsName = bucketFsName;
-            return this;
-        }
-
-        /**
-         * Set the bucket name.
-         *
-         * @param bucketName name of the bucket
-         * @return Builder instance for fluent programming
-         */
-        public Builder name(final String bucketName) {
-            this.bucketName = bucketName;
-            return this;
-        }
-
-        /**
-         * Set the IP address of the BucketFS service.
-         *
-         * @param ipAddress IP Address of the BucketFS service
-         * @return Builder instance for fluent programming
-         */
-        public Builder ipAddress(final String ipAddress) {
-            this.ipAddress = ipAddress;
-            return this;
-        }
-
-        /**
-         * Set the port the BucketFS service listens on.
-         *
-         * @param port HTTP port the BucketFS service listens on
-         * @return Builder instance for fluent programming
-         */
-        public Builder httpPort(final int port) {
-            this.port = port;
-            return this;
-        }
-
-        /**
-         * Set the read password.
-         *
-         * @param readPassword read password to set
-         * @return Builder instance for fluent programming
-         */
-        public Builder readPassword(final String readPassword) {
-            this.readPassword = readPassword;
-            return this;
-        }
-
-        /**
-         * Set the write password.
-         *
-         * @param writePassword write password to set
-         * @return Builder instance for fluent programming
-         */
-        public Builder writePassword(final String writePassword) {
-            this.writePassword = writePassword;
-            return this;
-        }
-
-        /**
-         * Build a new {@link Bucket} instance.
-         *
-         * @return bucket instance
-         */
-        public Bucket build() {
-            return new Bucket(this);
-        }
-    }
 }
