+++ src/main/java/com/exasol/bucketfs/BucketFactory.java
@@ -11,5 +11,5 @@ public interface BucketFactory {
      * @param bucketName  name of the bucket
      * @return bucket
      */
-    public Bucket getBucket(final String serviceName, final String bucketName);
+    public WriteEnabledBucket getBucket(final String serviceName, final String bucketName);
 }
