+++ src/main/java/com/exasol/bucketfs/ClusterConfigurationBucketFactory.java
@@ -10,7 +10,7 @@ import com.exasol.config.BucketFsServiceConfiguration;
  * Factory for objects abstracting buckets in Exasol's BucketFS.
  */
 public final class ClusterConfigurationBucketFactory implements BucketFactory {
-    private final Map<String, Bucket> bucketsCache = new HashMap<>();
+    private final Map<String, WriteEnabledBucket> bucketsCache = new HashMap<>();
     private final String ipAddress;
     private final BucketFsSerivceConfigurationProvider serviceConfigurationProvider;
     private final Map<Integer, Integer> portMappings;
@@ -39,7 +39,7 @@ public final class ClusterConfigurationBucketFactory implements BucketFactory {
 
     // [impl->dsn~bucket-factory-injects-access-credentials~1]
     @Override
-    public synchronized Bucket getBucket(final String serviceName, final String bucketName) {
+    public synchronized WriteEnabledBucket getBucket(final String serviceName, final String bucketName) {
         final String cacheKey = getFullyQualifiedBucketName(serviceName, bucketName);
         updateBucketCache(serviceName, bucketName, cacheKey);
         return getBucketFromCache(cacheKey);
@@ -49,7 +49,7 @@ public final class ClusterConfigurationBucketFactory implements BucketFactory {
         return serviceName + BucketConstants.PATH_SEPARATOR + bucketName;
     }
 
-    public Bucket getBucketFromCache(final String cacheKey) {
+    public WriteEnabledBucket getBucketFromCache(final String cacheKey) {
         return this.bucketsCache.get(cacheKey);
     }
 
@@ -57,7 +57,7 @@ public final class ClusterConfigurationBucketFactory implements BucketFactory {
         final BucketFsServiceConfiguration serviceConfiguration = this.serviceConfigurationProvider
                 .getBucketFsServiceConfiguration(serviceName);
         final BucketConfiguration bucketConfiguration = serviceConfiguration.getBucketConfiguration(bucketName);
-        this.bucketsCache.computeIfAbsent(cacheKey, bucket -> Bucket //
+        this.bucketsCache.computeIfAbsent(cacheKey, bucket -> WriteEnabledBucket //
                 .builder() //
                 .monitor(this.monitor) //
                 .serviceName(serviceName) //
