+++ src/test/java/com/exasol/bucketfs/AbstractBucketIT.java
@@ -17,22 +17,24 @@ abstract class AbstractBucketIT {
     protected static final ExasolContainer<? extends ExasolContainer<?>> EXASOL = new ExasolContainer<>() //
             .withReuse(true);
 
-    protected Bucket getDefaultBucket() {
-        final BucketFsServiceConfiguration serviceConfiguration = EXASOL.getClusterConfiguration() //
-                .getBucketFsServiceConfiguration(DEFAULT_BUCKETFS);
-        return getBucket(DEFAULT_BUCKET, serviceConfiguration);
+    protected String getContainerIpAddress() {
+        return EXASOL.getContainerIpAddress();
     }
 
-    private Bucket getBucket(final String name, final BucketFsServiceConfiguration serviceConfiguration) {
-        final BucketConfiguration bucketConfiguration = serviceConfiguration.getBucketConfiguration(name);
-        return Bucket.builder()//
-                .ipAddress(EXASOL.getContainerIpAddress()) //
-                .httpPort(EXASOL.getMappedPort(EXASOL.getDefaultInternalBucketfsPort())) //
-                .serviceName(serviceConfiguration.getName()) //
-                .name(name) //
-                .readPassword(bucketConfiguration.getReadPassword()) //
-                .writePassword(bucketConfiguration.getWritePassword()) //
-                .monitor(new LogBasedBucketFsMonitor(new LogPatternDetectorFactory(EXASOL))) //
-                .build();
+    protected Integer getMappedDefaultBucketFsPort() {
+        return EXASOL.getMappedPort(EXASOL.getDefaultInternalBucketfsPort());
+    }
+
+    protected BucketConfiguration getDefaultBucketConfiguration() {
+        return getDefaultBucketFsServiceConfiguration().getBucketConfiguration(DEFAULT_BUCKET);
+    }
+
+    private BucketFsServiceConfiguration getDefaultBucketFsServiceConfiguration() {
+        return EXASOL.getClusterConfiguration() //
+                .getBucketFsServiceConfiguration(DEFAULT_BUCKETFS);
+    }
+
+    protected LogBasedBucketFsMonitor createBucketMonitor() {
+        return new LogBasedBucketFsMonitor(new LogPatternDetectorFactory(EXASOL));
     }
 }
